Moved functionality into separate files for common, manager related, and server related
